# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenRelay API",
    "description": "Form orchestration backend.",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "me",
        "responses": {
          "200": {
            "description": "Current session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/oauth/admin-config": {
      "get": {
        "tags": [
          "oauth"
        ],
        "operationId": "admin_get_config",
        "responses": {
          "200": {
            "description": "Active OAuth config (no client_secret)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthConfigDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Insufficient permission"
          },
          "404": {
            "description": "No active config"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "oauth"
        ],
        "operationId": "admin_upsert_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertOAuthConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OAuth config saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthConfigDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Insufficient permission"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "oauth"
        ],
        "operationId": "admin_delete_config",
        "responses": {
          "204": {
            "description": "OAuth config removed"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Insufficient permission"
          },
          "404": {
            "description": "No active config"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/auth/oauth/discover": {
      "post": {
        "tags": [
          "oauth"
        ],
        "operationId": "admin_discover",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoveryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Discovery doc parsed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryPrefill"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Insufficient permission"
          },
          "502": {
            "description": "Discovery endpoint failed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/auth/password": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "change_own_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeOwnPassword"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Password changed; other sessions revoked"
          },
          "400": {
            "description": "Validation failed / no local password"
          },
          "401": {
            "description": "Missing/invalid token or wrong current password"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangeOwnPassword": {
        "type": "object",
        "description": "Self-service password change. Requires proof of the current password and\nrevokes the user's other sessions. See `service::change_own_password`.",
        "required": [
          "current_password",
          "new_password"
        ],
        "properties": {
          "current_password": {
            "type": "string"
          },
          "new_password": {
            "type": "string"
          }
        }
      },
...
      "DiscoveryPrefill": {
        "type": "object",
        "required": [
          "issuer",
          "authorize_url",
          "token_url"
        ],
        "properties": {
          "authorize_url": {
            "type": "string"
          },
          "issuer": {
            "type": "string"
          },
          "jwks_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "scopes_supported": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "token_url": {
            "type": "string"
          },
          "userinfo_url": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "DiscoveryRequest": {
        "type": "object",
        "required": [
          "discovery_url"
        ],
        "properties": {
          "discovery_url": {
            "type": "string"
          }
        }
      },
...
      "MeResponse": {
        "type": "object",
        "description": "Session-shape response for `/auth/me`. Flat permission set is what the\nfrontend's `usePermissions` hook consumes; `roles` provides the role\nbadges in the UI. Refresh on window focus so an admin's permission\nchanges propagate without forcing a re-login.",
        "required": [
          "user",
          "permissions",
          "roles"
        ],
        "properties": {
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleSummary"
            }
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        }
      },
...
      "OAuthConfigDto": {
        "type": "object",
        "description": "Admin-facing config — same shape minus `client_secret`. The admin UI\nshows `[unchanged]` for the secret field and only sends a value when the\nadmin types one.",
        "required": [
          "id",
          "display_name",
          "client_id",
          "authorize_url",
          "token_url",
          "scopes",
          "email_claim",
          "subject_claim",
          "has_client_secret"
        ],
        "properties": {
          "authorize_url": {
            "type": "string"
          },
          "client_id": {
            "type": "string"
          },
          "default_role_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "discovery_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "display_name": {
            "type": "string"
          },
          "email_claim": {
            "type": "string"
          },
          "has_client_secret": {
            "type": "boolean",
            "description": "True if a non-empty client_secret is on record."
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "issuer": {
            "type": [
              "string",
              "null"
            ]
          },
          "jwks_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "scopes": {
            "type": "string"
          },
          "subject_claim": {
            "type": "string"
          },
          "token_url": {
            "type": "string"
          },
          "userinfo_url": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Permission": {
        "type": "string",
        "enum": [
          "users:read",
          "users:write",
          "users:delete",
          "roles:read",
          "roles:write",
          "roles:delete",
          "roles:assign",
          "forms:read",
          "forms:write",
          "forms:delete",
          "submissions:read",
          "submissions:delete",
          "backends:read",
          "backends:write",
          "backends:delete",
          "auth_config:write"
        ]
      },
...
      "RoleSummary": {
        "type": "object",
        "description": "Lightweight role reference — enough to render a badge, populate a\ndropdown, or describe the current user's roles in `/auth/me`. Cheap to\nembed in `UserDto` since no permission join is required.",
        "required": [
          "id",
          "name",
          "is_system"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "is_system": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "UpsertOAuthConfig": {
        "type": "object",
        "description": "Upsert input. `client_secret: None` means \"keep the existing value\";\n`Some(non_empty)` replaces. The service rejects creating a new config\nwithout a client_secret.",
        "required": [
          "display_name",
          "client_id",
          "authorize_url",
          "token_url",
          "scopes"
        ],
        "properties": {
          "authorize_url": {
            "type": "string"
          },
          "client_id": {
            "type": "string"
          },
          "client_secret": {
            "type": [
              "string",
              "null"
            ]
          },
          "default_role_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "discovery_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "display_name": {
            "type": "string"
          },
          "email_claim": {
            "type": [
              "string",
              "null"
            ]
          },
          "issuer": {
            "type": [
              "string",
              "null"
            ]
          },
          "jwks_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "scopes": {
            "type": "string"
          },
          "subject_claim": {
            "type": [
              "string",
              "null"
            ]
          },
          "token_url": {
            "type": "string"
          },
          "userinfo_url": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "UserDto": {
        "type": "object",
        "description": "Outbound representation of a user — what callers see in API responses.\n\n`roles` is populated by `service::populate_roles` (or by the RBAC service\ndirectly for single-user lookups). `From<user::Model>` produces an empty\nroles vec; the responsibility to enrich falls to the caller so this type\nstays cheap to construct in places (login, setup) where the frontend will\nrefetch its session shape via `/auth/me` anyway.",
        "required": [
          "id",
          "email"
        ],
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleSummary"
            }
          }
        }
      },
...
    }
  },
...
}